Create Plotting Histogram Of An Image
The histogram of a digital image with gray levels in the range [0, L-1] is a discrete function. 
Histogram of an image provides a global description of the appearance of an image. Information obtained from histogram is very large in quality. Histogram of an image represents the relative frequency of occurrence of various gray levels in an image.