Create Histogram Equalization of Image
The histogram of a digital image, with intensity levels between 0 and (L-1), is a function h( rk ) = nk , where rk is the kth intensity level and nk is the number of pixels in the image having that intensity level. We can also normalize the histogram by dividing it by the total number of pixels in the image. 
For an N x N image, we have the following definition of a normalized histogram function: 

p( rk ) = nk/N2

This p(rk) function is the probability of the occurrence of a pixel with the intensity level rk. Clearly, 

∑ p( rk ) = 1

The histogram of an image, as shown in the figure, consists of the x-axis representing the intensity levels rk and the y-axis denoting the h(rk) or the p(rk) functions. 

The histogram of an image gives important information about the grayscale and contrast of the image. If the entire histogram of an image is centered towards the left end of the x-axis, then it implies a dark image. If the histogram is more inclined towards the right end, it signifies a white or bright image. 

In image processing, there frequently arises the need to improve the contrast of the image. In such cases, we use an intensity transformation technique known as histogram equalization.  Histogram equalization is the process of uniformly distributing the image histogram over the entire intensity axis by choosing a proper intensity transformation function. 
Hence, histogram equalization is an intensity transformation process.