Create Spatial Smoothing Filter - Mean Filter
Filtering techniques are use to enhance and modify digital images. Also, images filters are use to blurring and noise reduction , sharpening and edge detection.

Image filters can be divided in to two main categories. 
Spatial filtering is the traditional method of image filtering. it is use directly on the image pixels. 
Frequency domain filters are use to remove high and low frequencies and smoothing.

In Spatial Filtering technique mask is usually considered to be added in size so that it has specific center pixel. This mask is moved on the image such that the center of the mask traverses all image pixels.

There are two types of spatial filters- smoothing and sharpening spatial filters. 

Mean Filter- used for smoothing.
Linear spatial filter is simply the average of the pixels contained in the neighborhood of the filter mask. 
The idea is replacing the value of every pixel in an image by the average of the grey levels in the neighborhood define by the filter mask.

Types of Mean filter:

(i) Averaging filter: It is used in reduction of the detail in image. All coefficients are equal.
(ii) Weighted averaging filter: In this, pixels are multiplied by different coefficients. Center pixel is multiplied by a higher value than average filter.