Create Spatial Smoothing Filter- Median Filter
Order Statistics Filter: Type of smoothing spatial filter
It is based on the ordering the pixels contained in the image area encompassed by the filter. It replaces the value of the center pixel with the value determined by the ranking result. Edges are better preserved in this filtering.

Median filter: 
Each pixel in the image is considered. First neighboring pixels are sorted and original values of the pixel is replaced by the median of the list.